requirements.txt
selenium==4.15.2
requests==2.31.0
fake-useragent==1.4.0
webdriver-manager==4.0.1
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0 